+++ CHANGELOG.md
@@ -5,8 +5,9 @@ All notable changes to `Laravel Authentication Log` will be documented in this f
 ### 4.0.0 - 2024-03-28
 
 - Laravel 11 Support (https://github.com/rappasoft/laravel-authentication-log/pull/100)
-- Add config listeners
+- Add config listeners (https://github.com/rappasoft/laravel-authentication-log/pull/92)
 - Use real user IP behind Cloudflare
+- Check for AuthenticationLoggable trait on event (https://github.com/rappasoft/laravel-authentication-log/pull/94)
 
 ### 3.0.0 - 2023-02-23